Organization, moved AbstractShitBlock to common.
