+++ src/client/java/com/spirit/shit/block/custom/plush/KingZharaPlushBlock.java
@@ -1,5 +1,6 @@
 package com.spirit.shit.block.custom.plush;
 
+import com.spirit.shit.common.AbstractShitBlock;
 import net.minecraft.block.Block;
 import net.minecraft.util.function.BooleanBiFunction;
 import net.minecraft.util.shape.VoxelShape;
